Moved `girl_character_animations.lua` into `phase1` directory
