+++ prototypes/phase1/girl_character_animations.lua
